Add files via upload
In this work, we fix a choiced $M$, which is arbitrary, and we seek to model $\alpha$ as a function or as functions of volatility $\sigma$. We will see that accuracy is obtained when we consider put and call options separately. The purpose of these functions is to approximate the $\alpha$ values of any given volatility, and to interpolate the option price. We will see that this method works well to five decimal places of accuracy for the test volatility values $\sigma$, which are up to three decimal places in variation. All these ideas will be made clear and precise in the work proper below.